#16-refactoring: moved player classes to separate files
