+++ SoG_SGreader/Objects/Card.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class Card
+    {
+        public int CardID { get; set; }
+
+        public Card(int cardId)
+        {
+            this.CardID = cardId;
+        }
+    }
+}
