+++ SoG_SGreader/Objects/Enemy.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class Enemy
+    {
+        public int EnemyID { get; set; }
+
+        public Enemy(int enemyId)
+        {
+            this.EnemyID = enemyId;
+        }
+    }
+}
