+++ SoG_SGreader/Objects/Equip.cs
@@ -1,0 +1,14 @@
+namespace SoG_SGreader
+{
+    public class Equip
+    {
+        public int Hat { get; set; }
+        public int Facegear { get; set; }
+        public int Weapon { get; set; }
+        public int Shield { get; set; }
+        public int Armor { get; set; }
+        public int Shoes { get; set; }
+        public int Accessory1 { get; set; }
+        public int Accessory2 { get; set; }
+    }
+}
