+++ SoG_SGreader/Objects/FishCaught.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class FishCaught
+    {
+        public SogItems FishID { get; set; }
+
+        public FishCaught(SogItems fishId)
+        {
+            this.FishID = fishId;
+        }
+    }
+}
