+++ SoG_SGreader/Objects/Flag.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace SoG_SGreader
+{
+    public class Flag
+    {
+        public UInt16 FlagID { get; set; }
+
+        public Flag(UInt16 flagId) 
+        {
+            this.FlagID = flagId;
+        }
+    }
+}
