+++ SoG_SGreader/Objects/ItemCrafted.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class ItemCrafted
+    {
+        public SogItems ItemID { get; set; }
+
+        public ItemCrafted(SogItems itemId)
+        {
+            this.ItemID = itemId;
+        }
+    }
+}
