+++ SoG_SGreader/Objects/MerchantItem.cs
@@ -1,0 +1,13 @@
+namespace SoG_SGreader
+{
+    public class MerchantItem
+    {
+        public MerchantItem(SogItems _ItemID, int _ItemCount)
+        {
+            this.ItemID = _ItemID;
+            this.ItemCount = _ItemCount;
+        }
+        public SogItems ItemID { get; set; }
+        public int ItemCount { get; set; }
+    }
+}
