+++ SoG_SGreader/Objects/Player.ItemsSeen.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class ItemsSeen
+    {
+        public SogItems ItemID { get; set; }
+
+        public ItemsSeen(SogItems itemId)
+        {
+            this.ItemID = itemId;
+        }
+    }
+}
