+++ SoG_SGreader/Objects/Potion.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class Potion
+    {
+        public int PotionID { get; set; }
+
+        public Potion(int potionId)
+        {
+            this.PotionID = potionId;
+        }
+    }
+}
