+++ SoG_SGreader/Objects/Style.cs
@@ -1,0 +1,20 @@
+namespace SoG_SGreader
+{
+    public class Style
+    {
+        public char Bodytype { get; set; }
+        public int Hair { get; set; }
+        public int Hat { get; set; }
+        public int Facegear { get; set; }
+        public int Weapon { get; set; }
+        public int Shield { get; set; }
+        public bool HatHidden { get; set; }
+        public bool FacegearHidden { get; set; }
+        public int HairColor { get; set; }
+        public int SkinColor { get; set; }
+        public int PonchoColor { get; set; }
+        public int ShirtColor { get; set; }
+        public int PantsColor { get; set; }
+        public int Sex { get; set; }
+    }
+}
