+++ SoG_SGreader/Objects/TreasureMap.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace SoG_SGreader
+{
+    public class TreasureMap
+    {
+        public Int16 TreasureMapID { get; set; }
+
+        public TreasureMap(Int16 treasureMapID)
+        {
+            this.TreasureMapID = treasureMapID;
+        }
+    }
+}
