+++ SoG_SGreader/Objects/UnknownVariable01.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace SoG_SGreader
+{
+    public class UnknownVariable01
+    {
+        public Int16 UnknownVariable01ID { get; set; }
+
+        public UnknownVariable01(Int16 unknownVariable01Id)
+        {
+            this.UnknownVariable01ID = unknownVariable01Id;
+        }
+    }
+}
