+++ SoG_SGreader/Objects/UnknownVariable02.cs
@@ -1,0 +1,12 @@
+namespace SoG_SGreader
+{
+    public class UnknownVariable02
+    {
+        public byte UnknownByte { get; set; }
+
+        public UnknownVariable02(byte unknownByte)
+        {
+            this.UnknownByte = unknownByte;
+        }
+    }
+}
