+++ SoG_SGreader/Objects/UnknownVariable03.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace SoG_SGreader
+{
+    public class UnknownVariable03
+    {
+        public UInt16 UnknownVariable { get; set; }
+
+        public UnknownVariable03(UInt16 unknownVariable)
+        {
+            this.UnknownVariable = unknownVariable;
+        }
+    }
+}
