+++ SoG_SGreader/Objects/UnknownVariable07.cs
@@ -1,0 +1,13 @@
+namespace SoG_SGreader
+{
+    public class UnknownVariable07    //.archie.....lmb...C
+    {
+        public UnknownVariable07(string unknownString, float unknownFloat)
+        {
+            this.UnknownString = unknownString;
+            this.UnknownFloat = unknownFloat;
+        }
+        public string UnknownString { get; set; }
+        public float UnknownFloat { get; set; }
+    }
+}
